bugfix/#19 display decimal balance

+++ store/userSlice.ts
@@ -63,7 +63,7 @@ export const userSlice = createSlice({
     setUser: (state, action: PayloadAction<User>) => {
       state.user = action.payload;
       if (typeof action.payload.accountBalance === 'string') {
-        state.user.accountBalance = parseInt(action.payload.accountBalance)
+        state.user.accountBalance = parseFloat(action.payload.accountBalance)
       } else {
         state.user.accountBalance = action.payload.accountBalance
       }
